Extract DocumentService interface
